Add files via upload
The file thcsppych8p1 is the text and commentary of the first page of the 8th chapter of the larger file.  It contains a description of the base 3 number system employed in building the symbols used in the book.  The book is intended as a compliment to the Yi Ching and probably prompted the interpretation of the Yi Ching as employing a base 2 number system.